unload-acpi-table: add binary file unload-acpi-table
Unfortunately can't unload ACPI table with the
rm -rf command. Keep running into operation not
permitted. Also was too time consumming and inefficient
to load table then reboot if I wanted to load updated
table. Create C source file to remove the table
with the rmdir functions as that's the function
the table responds to.

Signed-off-by: Vincent Davis Jr <vince@underview.tech>